refactor: Rearrange utility functions
- TileUtils: ExtractTileRange(), IsWithinTileRange()
- GeometryUtils: SampleElevation()
